safari-534.53-branch/LayoutTests/platform/mac/accessibility/aria-tree-activedescendant-expected.txt

Animals
Birds
Cats
Siamese
Tabby
This tests that the ARIA drag and drop attributes work as intended.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS axtree.selectedRowAtIndex(0).isEqual(activeDescendant) is true
PASS axtree.selectedRowAtIndex(0).isEqual(newActiveDescendant) is true
PASS successfullyParsed is true

TEST COMPLETE
